requirements.txt

pillow
pickle
